Compression, first part
How to compress the generated 3D images ?
There are two big challenges:
* the bandwidth: 5 or 15 Mbit/s
* the onchip memory of FPGA: a bit less than 600 kbits

For the bandwidth, considering 20 fps, we have 250 kbits per frame in worst case.
For the onchip memory, due to double buffering and other uses of the onchip memory,
it would be good to not exceed 200 kbits per frame.

The screen is 800*480 = 384000 pixels. We need thus to compress at around 0.5 bit/pixel.

The compression scheme implemented is based on one fact:
the image is made of regions of constant color ^1.

When taking the representation of the image as a (1D) array of pixels
(encoded on 32 bits for instance). We have thus chunks of repeating values.

1. It is thus interesting to encode the data to (basic repetition coding)
  (Pixel color, Number of pixels with this color), ...
This works quite well (we have less than 10000 chunks in a typical image),
but is a bit short if we use a basic coding for the colors and chunk sizes.

2. We can also observe that that there are some pixel colors that have very
high probability (like background), and that there is a quite small amount of
colors (a few dozens). We will thus encode the pixel colors using
an Huffman code, and do the same for chunks lengths (that also have
interesting distributions)^2.

These methods combined give a size of around 40 kbits per frame.
(All the measurements are made for a simple picture: only
wall, structure and players, no global gauge, score, etc.)

On the implementation side, we generate a few sample images,
use it to construct Huffman trees (see python code), and
then generate the code for encoders/decoders. It is currently
done for C code, but the verilog decoder should not be much
more complex.
The only constraint used in implementation is that the size
of the codewords must be less than 32 bits.

^1 Actually, the lightning of the 3D scene is made to satisfy this property.
^2 For the chunk lengths, the set of possible values is much bigger than for
colors (think of large areas of background), so we need to add an upper limit.
This will allow to (with a few samples of images) correctly characterize
the distribution, and avoid large code size, which is annoying for
implementation of encoding/decoding.
This upper bound is not limiting: once it is reached, we start a new
chunk (with the same color).