Big refactoring of client.py
This opens the door to display more states:
the current state of the client is well-contained in an object
(client_core.GameState), which can be transmitted to the rendering
logic.

This also fixes a few bugs/missing features... I don't remember
the exact list.
One of them is the explicit action from the player to start the
game: it has to do a "pause" action to signal he is ready.

cc @anpar