Quantize colors to 4 bits, then add all colors to Huffman tree
This work quite well: we get maximum length of 18 bits for
Huffman color codes.

We should now be able to represent any image (up to quantization
and total size limits) with our compression scheme.

Current major drawback: compilation of compression.o take a bit
long (~20s on my laptop), it might take a few minutes to build
on a raspberry, and we need to do it each time we regenerate
Huffman trees (not too often, hopefully).
If build time is too long, we might try to change optimization level
to -O1 of -O0, at the expense of run time (uncertain how much...
benchmark needed).

A minor fix to verilog code was needed to handle codewords of size > 16